+++ constants/drops.js
@@ -14,3 +14,4 @@ export const MIDNIGHT_DYE = "Midnight Dye";
 export const PERIWINKLE_DYE = "Periwinkle Dye";
 export const BONE_DYE = "Bone Dye";
 export const MUSIC_RUNE = "Music Rune I";
+export const PHOENIX_PET = "Phoenix";